reworking the rendering system

+++ src/components.rs
@@ -4,8 +4,8 @@ use specs::VecStorage;
 #[derive(Debug, Component, Clone, Copy)]
 #[storage(VecStorage)]
 pub struct Position {
-    pub x: i64,
-    pub y: i64
+    pub x: u128,
+    pub y: u128
 }
 
 #[derive(Component)]
