+++ src/resources/next_iteration.rs
@@ -4,7 +4,7 @@ use crate::components::{LiveCell, DeadCell};
 
 #[derive(Default)]
 pub struct NextIteration {
-    pub live_cells: Vec<(i64, i64)>,
-    pub dead_cells: Vec<(i64, i64)>
+    pub live_cells: Vec<(u128, u128)>,
+    pub dead_cells: Vec<(u128, u128)>
 }
